View向けの関数をグローバルに移動
